+++ lib/ReactViewModels/ViewState.ts
@@ -73,6 +73,7 @@ export default class ViewState {
   @observable topElement: string = "FeatureInfo";
   @observable lastUploadedFiles: any[] = [];
   @observable storyBuilderShown: boolean = false;
+  @observable breadcrumbsShown: boolean = false;
 
   // Flesh out later
   @observable showHelpMenu: boolean = false;
@@ -510,6 +511,7 @@ export default class ViewState {
     this.explorerPanelIsVisible = false;
     this.switchMobileView(null);
     this.clearPreviewedItem();
+    this.showBreadcrumbs(false);
   }
 
   @action
@@ -565,9 +567,20 @@ export default class ViewState {
   }
 
   @action
+  changeSearchState(newText: string) {
+    this.searchState.catalogSearchText = newText;
+  }
+
+  @action
+  showBreadcrumbs(isActive: boolean) {
+    this.breadcrumbsShown = isActive;
+  }
+
+  @action
   setDisclaimerVisible(bool: boolean) {
     this.disclaimerVisible = bool;
   }
+
   @action
   hideDisclaimer() {
     this.setDisclaimerVisible(false);